+++ src/Nethermind/Nethermind.Evm/Instructions/EvmInstructions.Environment.cs
@@ -5,8 +5,9 @@ using System;
 using System.Diagnostics.CodeAnalysis;
 using System.Diagnostics;
 using System.Runtime.CompilerServices;
-using Nethermind.Core.Specs;
 using Nethermind.Core;
+using Nethermind.Core.Specs;
+using Nethermind.Core.Crypto;
 using Nethermind.Evm.EvmObjectFormat;
 using Nethermind.Evm.Precompiles;
 using Nethermind.State;
@@ -308,6 +309,13 @@ internal sealed partial class EvmInstructions
     /// <summary>
     /// Pushes the chain identifier onto the stack.
     /// </summary>
+    /// <returns>
+    /// <param name="vm">The virtual machine instance.</param>
+    /// <param name="stack">The execution stack.</param>
+    /// <param name="gasAvailable">The available gas which is reduced by the operation's cost.</param>
+    /// <param name="programCounter">The program counter.</param>
+    /// <see cref="EvmExceptionType.None"/>
+    /// </returns>
     [SkipLocalsInit]
     public static EvmExceptionType InstructionChainId(VirtualMachine vm, ref EvmStack stack, ref long gasAvailable, ref int programCounter)
     {
@@ -322,6 +330,15 @@ internal sealed partial class EvmInstructions
     /// Retrieves and pushes the balance of an account.
     /// The address is popped from the stack.
     /// </summary>
+    /// <param name="vm">The virtual machine instance.</param>
+    /// <param name="stack">The execution stack.</param>
+    /// <param name="gasAvailable">The available gas which is reduced by the operation's cost.</param>
+    /// <param name="programCounter">The program counter.</param>
+    /// <returns>
+    /// <see cref="EvmExceptionType.None"/> if gas is available,
+    /// <see cref="EvmExceptionType.OutOfGas"/> if the gas becomes negative
+    /// or <see cref="EvmExceptionType.StackUnderflow"/> if not enough items on stack.
+    /// </returns>
     [SkipLocalsInit]
     public static EvmExceptionType InstructionBalance(VirtualMachine vm, ref EvmStack stack, ref long gasAvailable, ref int programCounter)
     {
@@ -349,6 +366,13 @@ internal sealed partial class EvmInstructions
     /// <summary>
     /// Pushes the balance of the executing account onto the stack.
     /// </summary>
+    /// <param name="vm">The virtual machine instance.</param>
+    /// <param name="stack">The execution stack.</param>
+    /// <param name="gasAvailable">The available gas which is reduced by the operation's cost.</param>
+    /// <param name="programCounter">The program counter.</param>
+    /// <returns>
+    /// <see cref="EvmExceptionType.None"/>
+    /// </returns>
     [SkipLocalsInit]
     public static EvmExceptionType InstructionSelfBalance(VirtualMachine vm, ref EvmStack stack, ref long gasAvailable, ref int programCounter)
     {
@@ -365,6 +389,15 @@ internal sealed partial class EvmInstructions
     /// Retrieves the code hash of an external account.
     /// Returns zero if the account does not exist or is considered dead.
     /// </summary>
+    /// <param name="vm">The virtual machine instance.</param>
+    /// <param name="stack">The execution stack.</param>
+    /// <param name="gasAvailable">The available gas which is reduced by the operation's cost.</param>
+    /// <param name="programCounter">The program counter.</param>
+    /// <returns>
+    /// <see cref="EvmExceptionType.None"/> if gas is available,
+    /// <see cref="EvmExceptionType.OutOfGas"/> if the gas becomes negative
+    /// or <see cref="EvmExceptionType.StackUnderflow"/> if not enough items on stack.
+    /// </returns>
     [SkipLocalsInit]
     public static EvmExceptionType InstructionExtCodeHash(VirtualMachine vm, ref EvmStack stack, ref long gasAvailable, ref int programCounter)
     {
@@ -400,6 +433,15 @@ internal sealed partial class EvmInstructions
     /// Retrieves the code hash of an external account, considering the possibility of an EOF-validated contract.
     /// If the code is an EOF contract, a predefined EOF hash is pushed.
     /// </summary>
+    /// <param name="vm">The virtual machine instance.</param>
+    /// <param name="stack">The execution stack where the gas value will be pushed.</param>
+    /// <param name="gasAvailable">Reference to the current available gas, which is modified by this operation.</param>
+    /// <param name="programCounter">The current program counter.</param>
+    /// <returns>
+    /// <see cref="EvmExceptionType.None"/> if gas is available,
+    /// <see cref="EvmExceptionType.OutOfGas"/> if the gas becomes negative
+    /// or <see cref="EvmExceptionType.StackUnderflow"/> if not enough items on stack.
+    /// </returns>
     [SkipLocalsInit]
     public static EvmExceptionType InstructionExtCodeHashEof(VirtualMachine vm, ref EvmStack stack, ref long gasAvailable, ref int programCounter)
     {
@@ -442,6 +484,13 @@ internal sealed partial class EvmInstructions
     /// Implements the PREVRANDAO opcode.
     /// Pushes the previous random value (post-merge) or block difficulty (pre-merge) onto the stack.
     /// </summary>
+    /// <param name="vm">The virtual machine instance.</param>
+    /// <param name="stack">The execution stack.</param>
+    /// <param name="gasAvailable">The available gas which is reduced by the operation's cost.</param>
+    /// <param name="programCounter">The program counter.</param>
+    /// <returns>
+    /// <see cref="EvmExceptionType.None"/>
+    /// </returns>
     [SkipLocalsInit]
     public static EvmExceptionType InstructionPrevRandao(VirtualMachine vm, ref EvmStack stack, ref long gasAvailable, ref int programCounter)
     {
@@ -464,64 +513,119 @@ internal sealed partial class EvmInstructions
     }
 
     /// <summary>
-    /// Charges gas for accessing an account, including potential delegation lookups.
-    /// This method ensures that both the requested account and its delegated account (if any) are properly charged.
+    /// Pushes the remaining gas onto the stack.
+    /// The gas available is decremented by the base cost, and if negative, an OutOfGas error is returned.
     /// </summary>
-    /// <param name="gasAvailable">Reference to the available gas which will be updated.</param>
     /// <param name="vm">The virtual machine instance.</param>
-    /// <param name="address">The target account address.</param>
-    /// <param name="chargeForWarm">If true, charge even if the account is already warm.</param>
-    /// <returns>True if gas was successfully charged; otherwise false.</returns>
-    private static bool ChargeAccountAccessGasWithDelegation(ref long gasAvailable, VirtualMachine vm, Address address, bool chargeForWarm = true)
+    /// <param name="stack">The execution stack where the gas value will be pushed.</param>
+    /// <param name="gasAvailable">Reference to the current available gas, which is modified by this operation.</param>
+    /// <param name="programCounter">The current program counter.</param>
+    /// <returns>
+    /// <see cref="EvmExceptionType.None"/> if gas is available, or <see cref="EvmExceptionType.OutOfGas"/> if the gas becomes negative.
+    /// </returns>
+    [SkipLocalsInit]
+    public static EvmExceptionType InstructionGas(VirtualMachine vm, ref EvmStack stack, ref long gasAvailable, ref int programCounter)
     {
-        IReleaseSpec spec = vm.Spec;
-        if (!spec.UseHotAndColdStorage)
-        {
-            // No extra cost if hot/cold storage is not used.
-            return true;
-        }
-        bool notOutOfGas = ChargeAccountAccessGas(ref gasAvailable, vm, address, chargeForWarm);
-        return notOutOfGas
-               && (!vm.EvmState.Env.TxExecutionContext.CodeInfoRepository.TryGetDelegation(vm.WorldState, address, spec, out Address delegated)
-                   // Charge additional gas for the delegated account if it exists.
-                   || ChargeAccountAccessGas(ref gasAvailable, vm, delegated, chargeForWarm));
+        // Deduct the base gas cost for reading gas.
+        gasAvailable -= GasCostOf.Base;
+
+        // If gas falls below zero after cost deduction, signal out-of-gas error.
+        if (gasAvailable < 0) goto OutOfGas;
+
+        // Push the remaining gas (as unsigned 64-bit) onto the stack.
+        stack.PushUInt64((ulong)gasAvailable);
+
+        return EvmExceptionType.None;
+    // Jump forward to be unpredicted by the branch predictor.
+    OutOfGas:
+        return EvmExceptionType.OutOfGas;
     }
 
     /// <summary>
-    /// Charges gas for accessing an account based on its storage state (cold vs. warm).
-    /// Precompiles are treated as exceptions to the cold/warm gas charge.
+    /// Computes the blob hash from the provided blob versioned hashes.
+    /// Pops an index from the stack and uses it to select a blob hash from the versioned hashes array.
+    /// If the index is invalid, pushes zero.
     /// </summary>
-    /// <param name="gasAvailable">Reference to the available gas which will be updated.</param>
     /// <param name="vm">The virtual machine instance.</param>
-    /// <param name="address">The target account address.</param>
-    /// <param name="chargeForWarm">If true, applies the warm read gas cost even if the account is warm.</param>
-    /// <returns>True if the gas charge was successful; otherwise false.</returns>
-    public static bool ChargeAccountAccessGas(ref long gasAvailable, VirtualMachine vm, Address address, bool chargeForWarm = true)
+    /// <param name="stack">The execution stack from which the index is popped and where the blob hash is pushed.</param>
+    /// <param name="gasAvailable">Reference to the available gas; reduced by the blob hash cost.</param>
+    /// <param name="programCounter">The program counter.</param>
+    /// <returns>
+    /// <see cref="EvmExceptionType.None"/> on success; otherwise, <see cref="EvmExceptionType.StackUnderflow"/>
+    /// if there are insufficient elements on the stack.
+    /// </returns>
+    [SkipLocalsInit]
+    public static EvmExceptionType InstructionBlobHash(VirtualMachine vm, ref EvmStack stack, ref long gasAvailable, ref int programCounter)
     {
-        bool result = true;
         IReleaseSpec spec = vm.Spec;
-        if (spec.UseHotAndColdStorage)
-        {
-            EvmState vmState = vm.EvmState;
-            if (vm.TxTracer.IsTracingAccess)
-            {
-                // Ensure that tracing simulates access-list behavior.
-                vmState.AccessTracker.WarmUp(address);
-            }
 
-            // If the account is cold (and not a precompile), charge the cold access cost.
-            if (vmState.AccessTracker.IsCold(address) && !address.IsPrecompile(spec))
-            {
-                result = UpdateGas(GasCostOf.ColdAccountAccess, ref gasAvailable);
-                vmState.AccessTracker.WarmUp(address);
-            }
-            else if (chargeForWarm)
-            {
-                // Otherwise, if warm access should be charged, apply the warm read cost.
-                result = UpdateGas(GasCostOf.WarmStateRead, ref gasAvailable);
-            }
+        // Deduct the gas cost for blob hash operation.
+        gasAvailable -= GasCostOf.BlobHash;
+
+        // Pop the blob index from the stack.
+        if (!stack.PopUInt256(out UInt256 result)) goto StackUnderflow;
+
+        // Retrieve the array of versioned blob hashes from the execution context.
+        byte[][] versionedHashes = vm.EvmState.Env.TxExecutionContext.BlobVersionedHashes;
+
+        // If versioned hashes are available and the index is within range, push the corresponding blob hash.
+        // Otherwise, push zero.
+        if (versionedHashes is not null && result < versionedHashes.Length)
+        {
+            stack.PushBytes(versionedHashes[result.u0]);
+        }
+        else
+        {
+            stack.PushZero();
         }
 
-        return result;
+        return EvmExceptionType.None;
+    // Jump forward to be unpredicted by the branch predictor.
+    StackUnderflow:
+        return EvmExceptionType.StackUnderflow;
+    }
+
+    /// <summary>
+    /// Retrieves a block hash for a given block number.
+    /// Pops a block number from the stack, validates it, and then pushes the corresponding block hash.
+    /// If no valid block hash exists, pushes a zero value.
+    /// Additionally, reports the block hash if block hash tracing is enabled.
+    /// </summary>
+    /// <param name="vm">The virtual machine instance.</param>
+    /// <param name="stack">The execution stack from which the block number is popped and where the block hash is pushed.</param>
+    /// <param name="gasAvailable">Reference to the available gas; reduced by the block hash operation cost.</param>
+    /// <param name="programCounter">The program counter.</param>
+    /// <returns>
+    /// <see cref="EvmExceptionType.None"/> if the operation completes successfully;
+    /// otherwise, <see cref="EvmExceptionType.StackUnderflow"/> if there are insufficient stack elements.
+    /// </returns>
+    [SkipLocalsInit]
+    public static EvmExceptionType InstructionBlockHash(VirtualMachine vm, ref EvmStack stack, ref long gasAvailable, ref int programCounter)
+    {
+        // Deduct the gas cost for block hash operation.
+        gasAvailable -= GasCostOf.BlockHash;
+
+        // Pop the block number from the stack.
+        if (!stack.PopUInt256(out UInt256 a)) goto StackUnderflow;
+
+        // Convert the block number to a long. Clamp the value to long.MaxValue if it exceeds it.
+        long number = a > long.MaxValue ? long.MaxValue : (long)a.u0;
+
+        // Retrieve the block hash for the given block number.
+        Hash256? blockHash = vm.BlockHashProvider.GetBlockhash(vm.EvmState.Env.TxExecutionContext.BlockExecutionContext.Header, number);
+
+        // Push the block hash bytes if available; otherwise, push a 32-byte zero value.
+        stack.PushBytes(blockHash is not null ? blockHash.Bytes : BytesZero32);
+
+        // If block hash tracing is enabled and a valid block hash was obtained, report it.
+        if (vm.TxTracer.IsTracingBlockHash && blockHash is not null)
+        {
+            vm.TxTracer.ReportBlockHash(blockHash);
+        }
+
+        return EvmExceptionType.None;
+    // Jump forward to be unpredicted by the branch predictor.
+    StackUnderflow:
+        return EvmExceptionType.StackUnderflow;
     }
 }
